package/e2fsprogs: add dependency on host-coreutils
When performing a build on a host system without the coreutils
package (busybox based), e2fsprogs install step will fail with the
following error:

  making install in lib/et
          MKDIR_P /usr/lib /usr/lib
  make[3]: config/install-sh: No such file or directory
  make[3]: *** [Makefile:448: installdirs-elf-lib] Error 127

This is caused by using the busybox version of mkdir. Fix the error by
add dependency on host-coreutils for both host and target packages of
e2fsprogs.

Signed-off-by: Ryan Barnett <ryanbarnett3@gmail.com>